Refactor query out of domainList to use in mutation
