+++ src/components/Address/Address.js
@@ -1,5 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import styled from '@emotion/styled'
+import { useQuery } from 'react-apollo'
+
+import { GET_DOMAINS_OWNED_BY_ADDRESS_FROM_SUBGRAPH } from '../../graphql/queries'
 
 import mq from 'mediaQuery'
 
@@ -12,6 +15,7 @@ import DomainList from './DomainList'
 import RenewAll from './RenewAll'
 import Sorting from './Sorting'
 import Filtering from './Filtering'
+import Loader from '../Loader'
 
 const TopBar = styled(DefaultTopBar)`
   margin-bottom: 40px;
@@ -43,7 +47,62 @@ const Controls = styled('div')`
   `}
 `
 
+function filterOutReverse(domains) {
+  return domains.filter(
+    domain => domain.parent && domain.parent.name !== 'addr.reverse'
+  )
+}
+
+function normaliseAddress(address) {
+  return address.toLowerCase()
+}
+
+function getSortFunc(activeSort) {
+  function alphabetical(a, b) {
+    if (
+      a.domain.name &&
+      a.domain.name[0] === '[' &&
+      b.domain.name &&
+      b.domain.name[0] === '['
+    )
+      return a.domain.name > b.domain.name ? 1 : -1
+    if (a.domain.name && a.domain.name[0] === '[') return 1
+    if (b.domain.name && b.domain.name[0] === '[') return -1
+    return a.domain.name > b.domain.name ? 1 : -1
+  }
+  switch (activeSort) {
+    case 'alphabetical':
+      return alphabetical
+    case 'alphabeticalDesc':
+      return (a, b) => {
+        if (
+          a.domain.name &&
+          a.domain.name[0] === '[' &&
+          b.domain.name &&
+          b.domain.name[0] === '['
+        )
+          return a.domain.name > b.domain.name ? 1 : -1
+        if (a.domain.name && a.domain.name[0] === '[') return 1
+        if (b.domain.name && b.domain.name[0] === '[') return -1
+        return a.domain.name < b.domain.name ? 1 : -1
+      }
+    case 'expiryDate':
+      return (a, b) => a.expiryDate - b.expiryDate
+
+    case 'expiryDateDesc':
+      return (a, b) => b.expiryDate - a.expiryDate
+    default:
+      return alphabetical
+  }
+}
+
 export default function Address({ address }) {
+  const normalisedAddress = normaliseAddress(address)
+  const { loading, data, error } = useQuery(
+    GET_DOMAINS_OWNED_BY_ADDRESS_FROM_SUBGRAPH,
+    { variables: { id: normalisedAddress } }
+  )
+
   let [etherScanAddr, setEtherScanAddr] = useState(null)
   let [activeSort, setActiveSort] = useState('alphabetical')
   let [activeFilter, setActiveFilter] = useState('registrant')
@@ -53,6 +112,32 @@ export default function Address({ address }) {
   useEffect(() => {
     getEtherScanAddr().then(setEtherScanAddr)
   }, [])
+
+  if (error) {
+    return 'Error getting domains'
+  }
+
+  if (loading) {
+    return <Loader withWrap large />
+  }
+
+  let domains = []
+
+  if (activeFilter === 'registrant') {
+    domains = [...data.account.registrations?.sort(getSortFunc(activeSort))]
+  } else if (activeFilter === 'controller') {
+    domains = [
+      ...filterOutReverse(data.account.domains)
+        .map(domain => ({ domain }))
+        .sort(getSortFunc(activeSort))
+    ]
+  }
+
+  const selectedNames = Object.entries(checkedBoxes)
+    .filter(([key, value]) => value)
+    .map(([key]) => key)
+
+  const allNames = domains.map(d => d.domain.name)
 
   return (
     <AddressContainer>
@@ -82,14 +167,15 @@ export default function Address({ address }) {
         <RenewAll
           years={years}
           setYears={setYears}
-          selectedNames={Object.entries(checkedBoxes)
-            .filter(([key, value]) => value)
-            .map(([key]) => key)}
+          activeFilter={activeFilter}
+          selectedNames={selectedNames}
+          allNames={allNames}
         />
       </Controls>
 
       <DomainList
         address={address}
+        domains={domains}
         activeSort={activeSort}
         activeFilter={activeFilter}
         checkedBoxes={checkedBoxes}
